Add support for real tokens :)
- Upon login now a token is created in the database
- You can either pass a location parameter which will be validated to be
a valid location_id for your user's account, or if it is invalid or if
you don't pass a location then your authenticated user's
default_location will be used.
- This token will be used to provide the logged in user and that user's
location to other API calls
- This broke a test so I will fix that in the next checkin or so
- I also need to add this to the Rails Cache so that the token is cached
on creation and doesn't have to be constantly looked up in the database
on every request.